+++ huracanpy/__init__.py
@@ -1,4 +1,6 @@
-"""huracanpy - A python package for working with various forms of feature tracking data"""
+"""
+huracanpy - A python package for working with various forms of feature tracking data
+"""
 
 __all__ = [
     # Modules
@@ -13,6 +15,7 @@ __all__ = [
     "sel_id",
     "trackswhere",
     "interp_time",
+    "concat_tracks",
     # Parameters
     "basins",
     "example_csv_file",
@@ -28,6 +31,8 @@ __all__ = [
     "example_ERA20C_file",
     "example_WiTRACK_file",
     "example_old_HURDAT_file",
+    "example_STORM_file",
+    "example_IRIS_file",
     "_accessor",
 ]
 
@@ -47,9 +52,12 @@ from ._data import (
     example_ERA20C_file,
     example_WiTRACK_file,
     example_old_HURDAT_file,
+    example_STORM_file,
+    example_IRIS_file,
 )
 from ._basins import basins
 from ._interp import interp_time
+from ._concat import concat_tracks
 from ._subset import sel_id, trackswhere
 from . import calc, plot, assess, info, tc