+++ huracanpy/_data/_csv.py
@@ -28,30 +28,6 @@ pandas_na_values = [
     "",
 ]
 
-pandas_valid_time_labels = [
-    "year",
-    "years",
-    "month",
-    "months",
-    "day",
-    "days",
-    "hour",
-    "hours",
-    "minute",
-    "minutes",
-    "second",
-    "seconds",
-    "ms",
-    "millisecond",
-    "milliseconds",
-    "us",
-    "microsecond",
-    "microseconds",
-    "ns",
-    "nanosecond",
-    "nanoseconds",
-]
-
 
 def load(
     filename,
@@ -66,7 +42,8 @@ def load(
     filename : str
         The file must contain at least longitude, latitude, time and track ID.
             - longitude and latitude can be named that, or lon and lat.
-            - time must be defined a single `time`column or by four columns : year, month, day, hour
+            - time must be defined a single `time`column or by four columns:
+              year, month, day, hour
             - track ID must be within a column named track_id.
 
     load_function : callable
@@ -89,34 +66,7 @@ def load(
     # Remove leading/trailing spaces and make all column names lowercase
     tracks.columns = tracks.columns.str.strip().str.lower()
 
-    ## Time attribute
-    if "iso_time" in tracks.columns:
-        tracks["time"] = pd.to_datetime(tracks.iso_time)
-        tracks = tracks.drop(columns="iso_time")
-    elif "isotime" in tracks.columns:
-        tracks["time"] = pd.to_datetime(tracks.isotime)
-        tracks = tracks.drop(columns="isotime")
-    elif "time" in tracks.columns:
-        tracks["time"] = pd.to_datetime(tracks.time)
-    else:
-        # Combine separate year/month/day etc. values into a time, and drop those
-        # variables from the dataframe
-        time_vars = [var for var in tracks.columns if var in pandas_valid_time_labels]
-        tracks["time"] = pd.to_datetime(tracks[time_vars])
-        tracks = tracks.drop(time_vars, axis="columns")
-
     # Output xr dataset
     tracks = tracks.to_xarray().rename({"index": "record"}).drop_vars("record")
-
-    # Convert any variables that are objects to explicit string objects.
-    # This can cause strings stored as "NA", such as for basin to be converted to NaNs
-    # Revert any nans back to their original values
-    # Better to do it here than getting caught out later
-    for var in tracks:
-        if tracks[var].dtype == "O" and isinstance(tracks[var].data[0], str):
-            new_var = tracks[var].astype(str)
-            false_nans = new_var == "nan"
-            new_var[false_nans] = tracks[var][false_nans]
-            tracks[var] = new_var
 
     return tracks